Updates for Svc/ActiveLogger corresponding to UT cmake udpate for protected/private
